pyproject.toml
[tool.poetry]
name = "copylibs"
version = "0.1.0"
description = "Recursively identify libraries needed for a Python project"
authors = ["Lance Jacobs <lance@ljacobs.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.4"
argparse = "^1.4.0"
colorama = "^0.4.6"
frosch = "^0.1.9"
pyelftools = "^0.29"

[tool.poetry.scripts]
copylibs = "copylibs.copylibs:main"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
